add track assignment modal
